[feat]:리뷰생성,삭제 이벤트 발행 -> 포인트 생성,삭제 테스트(Kafka역직렬화 문제 해결 및 테스트 완료) #72

+++ benefit-service/src/main/java/com/waitless/benefit/point/infrastructure/adaptor/in/messaging/ReviewDeletedEventHandler.java
@@ -5,6 +5,9 @@ import com.waitless.common.event.ReviewDeletedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.support.KafkaHeaders;
+import org.springframework.messaging.handler.annotation.Header;
+import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Component;
 
 @Slf4j
@@ -15,13 +18,15 @@ public class ReviewDeletedEventHandler {
     private final PointCommandUseCase pointCommandUseCase;
 
     @KafkaListener(
-            topics = "review-events",
-            groupId = "point-service",
-            properties = {
-                    "spring.json.value.default.type=com.waitless.common.event.ReviewDeletedEvent"
-            }
+            topics = "review-deleted-events",
+            groupId = "point-service"
     )
-    public void handleReviewDeletedEvent(ReviewDeletedEvent event) {
+    public void handleReviewDeletedEvent(
+            @Payload ReviewDeletedEvent event,
+            @Header(KafkaHeaders.RECEIVED_KEY) String key
+    ) {
+        if (!"review-deleted".equals(key)) return;
+
         log.info("[Kafka] ReviewDeletedEvent 수신: {}", event);
         pointCommandUseCase.deletePointByReview(event.getReviewId(), event.getUserId());
         log.info("포인트 삭제 완료: reviewId={}", event.getReviewId());
